@@ -16,10 +16,16 @@
     "prod": "npm-run-all --parallel server api-server"
   },
   "dependencies": {
-    "@auth0/auth0-react": "^1.2.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.17",
-    "@fortawesome/free-solid-svg-icons": "^5.8.1",
-    "@fortawesome/react-fontawesome": "^0.1.4",
+    "@auth0/auth0-react": "^1.9.0",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.16",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/material": "^5.2.8",
+    "@mui/styled-engine-sc": "^5.1.0",
+    "atomize": "^1.0.28",
     "cors": "^2.8.5",
     "express": "^4.16.4",
     "express-jwt": "^6.0.0",
@@ -30,12 +36,16 @@
     "jwks-rsa": "^1.4.0",
     "morgan": "^1.9.1",
     "npm-run-all": "^4.1.5",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "prop-types": "^15.8.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "react-transition-group": "^4.4.2",
     "reactstrap": "^8.0.0",
+    "styled-components": "^5.3.3",
+    "styletron-engine-atomic": "^1.4.8",
+    "styletron-react": "^6.0.2",
     "web-vitals": "^2.1.3",
     "workbox-background-sync": "^5.1.4",
     "workbox-broadcast-update": "^5.1.4",
@@ -53,12 +63,12 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.10.1",
     "@testing-library/react": "^10.3.0",
-    "eslint-config-airbnb": "19.0.1",
+    "eslint-config-airbnb": "19.0.4",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
     "jest-environment-jsdom-sixteen": "^1.0.3",
     "nodemon": "^2.0.4",
@@ -82,7 +92,13 @@
             ".jsx"
           ]
         }
-      ]
+      ],
+      "react/no-unescaped-entities": 0,
+      "no-unused-vars": "warn",
+      "react/function-component-definition": [
+        0
+      ],
+      "react/display-name": 0
     }
   },
   "browserslist": {
