@@ -1,20 +1,16 @@
 import React from 'react';
-import { Router, Route, Switch } from 'react-router-dom';
-import { Container } from 'reactstrap';
+import { Router, Route, Routes } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
 
 import Loading from './components/Loading';
-import NavBar from './components/NavBar';
-import Footer from './components/Footer';
 import Home from './views/Home';
 import Profile from './views/Profile';
 import ExternalApi from './views/ExternalApi';
 import history from './utils/history';
 import initFontAwesome from './utils/initFontAwesome';
-
-// fontawesome
-// styles
+import Recipe from './views/Recipe';
 import './App.css';
+import FloatActionButton from './components/floatActionButton';
 
 initFontAwesome();
 
@@ -30,17 +26,15 @@ const App = function App() {
   }
 
   return (
-    <Router history={history}>
-      <div id="app" className="d-flex flex-column h-100">
-        <NavBar />
-        <Container className="flex-grow-1 mt-5">
-          <Switch>
-            <Route path="/" exact component={Home} />
-            <Route path="/profile" component={Profile} />
-            <Route path="/external-api" component={ExternalApi} />
-          </Switch>
-        </Container>
-        <Footer />
+    <Router location={history.location} navigator={history}>
+      <div id="app" style={{ backgroundColor: '#F1F8F6' }}>
+        <Routes>
+          <Route path="/" exact element={<Home />} />
+          <Route path="/recipes" element={<Recipe />} />
+          <Route path="/profile" element={<Profile />} />
+          <Route path="/external-api" element={<ExternalApi />} />
+        </Routes>
+        <FloatActionButton />
       </div>
     </Router>
   );
