@@ -4,7 +4,7 @@ import logo from '../assets/logo.svg';
 
 const Hero = function Hero() {
   return (
-    <div className="text-center hero my-5">
+    <div className="d-flex flex-column hero justify-content-center">
       <img className="mb-3 app-logo" src={logo} alt="React logo" width="120" />
       <h1 className="mb-4">React.js Sample Project</h1>
 
